Moved design system utility functions back for future refactoring and to fix tests using object key iterator
