+++ packages/fast-components-styles-msft/src/design-system/index.ts
@@ -2,7 +2,6 @@ import { Direction } from "@microsoft/fast-web-utilities";
 import { white } from "../utilities/color/color-constants";
 import { ColorPalette, ColorRGBA64, parseColorHexRGB } from "@microsoft/fast-colors";
 import { Palette } from "../utilities/color/palette";
-import { TypeRamp } from "../utilities/typography";
 import { withDefaults } from "@microsoft/fast-jss-utilities";
 import { defaultFontWeights, FontWeight } from "../utilities/fonts";
 
@@ -209,12 +208,37 @@ export const withDesignSystemDefaults: (config: Partial<DesignSystem>) => Design
 
 /**
  * Safely retrieves a single property from a design system
- * @deprecated Use getDesignSystemValue instead (utilities).
+ * @deprecated Use getDesignSystemValue instead.
  */
 export function getDesignSystemProperty(key: string): DesignSystemResolver<string> {
     return function(config: DesignSystem): string {
         return withDesignSystemDefaults(config)[key];
     };
+}
+
+/**
+ * Safely retrieves the value from a key of the designSystem.
+ */
+export function getDesignSystemValue<T extends DesignSystem, K extends keyof T>(
+    key: K
+): (designSystem?: T) => T[K] {
+    return (designSystem?: T): T[K] =>
+        (designSystem && designSystem[key]) || (designSystemDefaults as T)[key];
+}
+
+/**
+ * Returns the argument if basic, othwerwise calls the DesignSystemResolver function.
+ *
+ * @param arg A value or a DesignSystemResolver function
+ * @param designSystem The design system config.
+ */
+export function checkDesignSystemResolver<T>(
+    arg: T | DesignSystemResolver<T>,
+    designSystem: DesignSystem
+): T {
+    const value: T =
+        typeof arg === "function" ? (arg as DesignSystemResolver<T>)(designSystem) : arg;
+    return value;
 }
 
 /**
